feat(sample): add higher-order lettable version of sample
